fix: bug with write + add sparkle

+++ main/include/artnet.h
@@ -5,8 +5,8 @@
 #define DMX_DATA_MAX_LENGTH 512
 
 int artnet_init(char * destination_ip);
-int artnet_send_dmx(uint16_t universe, const uint8_t *dmx_data, uint16_t dmx_data_length);
-int artnet_send_dmx_array(const uint8_t dmx_data_array[][DMX_DATA_MAX_LENGTH], uint16_t num_universes);
+int artnet_send_dmx(uint8_t universe, const uint8_t *dmx_data, uint16_t dmx_data_length);
+int artnet_send_dmx_array(const uint8_t dmx_data_array[][DMX_DATA_MAX_LENGTH], uint8_t num_universes);
 int artnet_close();
 #endif // ARTNET_H
